+++ my-react-app/src/kilter-board/loader.js
@@ -2,7 +2,7 @@ import React, { useState, useEffect, useMemo } from "react";
 import { useParams } from "react-router-dom";
 import KilterBoard from "./KilterBoard";
 import { fetchBoardDetails } from "../board-page/api";
-import { boardLayouts } from "../board-page/board-data";
+import { boardLayouts } from "./board-data";
 
 // TODO: Unhardcode set_ids
 const KilterBoardLoader = ({ litUpHolds }) => {
